+++ aequilibrae/paths/path_based/TrafficAssignment.h
@@ -90,7 +90,7 @@ public:
 	void update_link_flows(unsigned int from_node);
 	void update_link_flows_stepsize(double stepsize);
 	void update_link_flows_by_origin(unsigned int from_node);
-    void update_path_flows_stepsize(unsigned int origin, float stepsize);
+    void update_path_flows_stepsize(unsigned int origin, double stepsize);
 
 	void insert_od(unsigned long from, unsigned long to, float demand);
